--- a/lib/ivy.xml
+++ b/lib/ivy.xml
@@ -12,5 +12,6 @@
     <dependency org="net.sourceforge.collections" name="collections-generic" rev="4.01"/>
     <dependency org="args4j" name="args4j" rev="2.0.29" />
     <dependency org="com.google.code.gson" name="gson" rev="2.5"/>
+    <dependency org="com.google.guava" name="guava" rev="19.0"/>
   </dependencies>
 </ivy-module>