--- a/pom.xml
+++ b/pom.xml
@@ -122,5 +122,10 @@
       <version>4.10</version>
       <optional>true</optional>
     </dependency>
+    <dependency>
+        <groupId>com.google.guava</groupId>
+        <artifactId>guava</artifactId>
+        <version>19.0</version>
+    </dependency>
   </dependencies>
 </project>