--- a/resources/wa_grammar.packed/config
+++ b/resources/wa_grammar.packed/config
@@ -1,0 +1,1 @@
+max-source-len = 6