--- a/test/lattice-short/test.sh
+++ b/test/lattice-short/test.sh
@@ -1,6 +1,6 @@
 #!/bin/sh
 
-cat input | $JOSHUA/bin/joshua-decoder -m 500m -config joshua.config 2> log | sort > output
+cat input | $JOSHUA/bin/joshua-decoder -m 500m -config joshua.config 2> log > output
 
 if [[ $? -ne 0 ]]; then
 	exit 1