… resources/grammar.glue
… resources/wa_grammar
… resources/wa_grammar.packed/encoding
… resources/wa_grammar.packed/slice_00000.alignments
… resources/wa_grammar.packed/slice_00000.features
… resources/wa_grammar.packed/slice_00000.source
… resources/wa_grammar.packed/slice_00000.target
… resources/wa_grammar.packed/slice_00000.target.lookup
… resources/wa_grammar.packed/vocabulary
… src/joshua/decoder/ff/tm/packed/PackedGrammar.java
… src/joshua/decoder/hypergraph/AlignedSourceTokens.java
… src/joshua/decoder/hypergraph/ViterbiExtractor.java
… src/joshua/decoder/hypergraph/WordAlignmentExtractor.java
… src/joshua/decoder/hypergraph/WordAlignmentState.java
… test/bn-en/packed/output.scores.gold
… tst/joshua/system/AlignmentMapTest.java
… tst/joshua/system/MultithreadedTranslationTests.java
… tst/joshua/system/StructuredOutputTest.java
… tst/joshua/system/StructuredTranslationTest.java