Hello Joshua folks.  I've got a series of patches to contribute to Joshua including for the most part a lot of performance improvements.  

**Performance Improvements**
The two main changes in terms of performance are that we've removed some string parsing that was not strictly needed, and slowed decoding.  We've also added a post-deserialization LRU cache for objects that are commonly requested from the packed grammar during decoding.  Because the same objects are requested over and over again from the packed grammar this reduces the cost of unpacking them, allocating new objects, calculating their feature weights, etc.  It also dramatically reduces the work the garbage collector needs to do during decoding.  

**New Features**
In terms of new features we do add a small new feature that we will further develop in future commits.  Namely the ability to extract information via a StructuredTranslation object that provides similar information to what is displayed to stdout.  This means if you are using Joshua as a library, the consuming application will have programatic access to investing aspects to the translate call, for example what the alignments in this translation were.

**New Dependencies**
These initial few patches add a new dependency, Google Guava 19.
http://mvnrepository.com/artifact/com.google.guava/guava/19.0
https://github.com/google/guava
LICENSE: https://github.com/google/guava/blob/master/COPYING (Apache 2.0)

**Testing**
We've also got a number of unit tests that we'll be contributing.  In this pull request there's only two, but there should be more to come in further pulls.  Currently these tests don't have a build target, but we can add that in the future if you (the Joshua maintainers) would find them beneficial.