@mjpost 2016-04-02
This is awesome. I just pulled this in after verifying the regression tests passed (github.com/joshua-decoder/joshua is still the main until the switch happens, hopefully next week). I'll run my own timing requests and report them here soon.