@mjpost 2016-04-04
It's merged into the [github repo](https://github.com/joshua-decoder/joshua) (which I'm still using until we pull off the whole switch) and also to [the Apache git repo](https://git-wip-us.apache.org/repos/asf?p=incubator-joshua.git;a=summary).