@mjpost 2016-04-05
For reference, I ran this across my timing suite, running start to finish, single-threaded, iincluding model loading time). For two packed models:

Phrase-based es-en model, 3k sentences: 495 seconds → 368 seconds (1.3x faster)
Hiero zh-en model (1,357 sentences): 586 seconds → 259 seconds (2.2x faster)

Decoding with a RAM-loaded model (unpacked model) is the same. This merge basically means that packed grammars are no longer any slower than just loading to memory, and also don't pay the startup cost. This is obviously really awesome.